Split adapter tests by backend (#114)
Move the shared Python command builder into private adapter test support
and split the mixed adapter test module into logging, metrics, and tracing
modules.

Keep the existing adapter test bodies and names intact while removing the
mixed-responsibility catch-all module.
